requirements.txt
pandas
scikit-learn
tensorflow-gpu==2.10.0
opencv-contrib-python==4.9.0.80
matplotlib
numpy<2.0
albumentations
xgboost
imbalanced-learn
seaborn
scikit-image
tqdm
scikit-posthocs>=0.7
tensorflow-addons